Move files to their distinct directories
